+++ src/main/java/insane96mcp/iguanatweaksreborn/entity/ITRFallingBlockEntity.java
@@ -124,7 +124,7 @@ public class ITRFallingBlockEntity extends FallingBlockEntity {
 		boolean maxStackReached = false;
 		while (true) {
 			blockPos.set(blockPos.above());
-			if (blockPos.getY() - pos.getY() > 8) {
+			if (blockPos.getY() - pos.getY() > 6) {
 				maxStackReached = true;
 				/*if (this.dropItem && this.level().getGameRules().getBoolean(GameRules.RULE_DOENTITYDROPS)) {
 					this.discard();